feat: disable fileWatcher by default (#225)
* fix: fix detection of changed files

* feat: queue filewatcher changes

This helps when many files are changed and prevents constant recompiling
as the cache is invalidated.

* feat: disable fileWatcher by default

FileWatcher is now under the Advanced Menu as a setting. This should
disable the code that may be killing bad Linux installs by gating the
use of `std::filesystem::last_write_time` and `esfw` behind a setting.